Fix intrinsic size for IndeterminateProgressDrawable; Other minor fixes.

+++ library/src/main/java/me/zhanghai/android/materialprogressbar/IndeterminateHorizontalProgressDrawable.java
@@ -21,7 +21,7 @@ public class IndeterminateHorizontalProgressDrawable extends IndeterminateProgre
         implements ShowTrackDrawable {
 
     private static final float PROGRESS_INTRINSIC_HEIGHT_DP = 3.2f;
-    private static final float PADDED_INTRINSIC_HEIGHT_DP = 16;
+    private static final int PADDED_INTRINSIC_HEIGHT_DP = 16;
     private static final RectF RECT_BOUND = new RectF(-180, -1, 180, 1);
     private static final RectF RECT_PADDED_BOUND = new RectF(-180, -5, 180, 5);
     private static final RectF RECT_PROGRESS = new RectF(-144, -1, 144, 1);
